Makefile.uk: Allow configuring with custom config file
The `.config.orig` rule does a backup of the config file before changing
it to set `CONFIG_STACK_SIZE_PAGE_ORDER=10`. This used a hardcoded path
to the default `.config` file instead of `$(C)`, which can be set to a
custom one.

Using a hardcoded path will fail if trying to build with a different
config file than the default one (i.e. using `make C=path-to-config`).
One of the usecases is `kraftkit`, which uses `.config.python3...`
instead of the default `.config` file.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>