Makefile.uk: Fix generated header include path
Previously LIBPYTHON3_SRC was in the include path, leading to a wrong
pyconfig.h, if present, to be pulled into the Python build. Such a file
could remain after being generated by a rootfs build.
This change both removes LIBPYTHON3_SRC from the include path, as it
provides no needed headers, as well as ensures that the rootfs build
does not leave a pyconfig.h file in the origin directory.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #19